S10P12A407-785 feat: 게시판 검색 기능 추가(게시판 목록 api 수정)

+++ backend/src/main/java/com/a407/back/dto/board/BoardListResponse.java
@@ -11,7 +11,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class BoardListResponse {
 
-    private Long totalCount;
+    private Integer totalCount;
     private int nowPage;
     private List<BoardListDto> boardList;
